# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Amazonet",
        "description": "This platform, similar to Amazon. It enables user authentication, product management, catalog browsing, shopping cart and checkout processes, order management, and more. The platform supports multiple roles, including customers, deliveryman, and administrators, providing each with specific features to manage products, orders, and user interactions."
    },
    "servers": [
        {
            "url": "http://localhost:4000/"
        }
    ],
...
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Login user",
                "description": "Login user",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Login user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "email@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "your_password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You have successfully logged in"
                                        },
                                        "userId": {
                                            "type": "string",
                                            "example": "66c4a618f0ad07493daf1e96"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "Role"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "refreshToken": {
                                            "type": "string",
                                            "example": "refresh token"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email or password!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid email or password!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Display one user",
                "description": "Display one user",
                "operationId": "displayOneUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found!"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "UpdateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Created user object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "new_name"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "new_password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully update the profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully update the profile"
                                        },
                                        "userId": {
                                            "type": "string",
                                            "example": "66b8db97e4b437c6b2569e79"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "description": "Delete user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The user that needs to be deleted",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Can not delete Admin User!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Can not delete Admin User!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "role"
                ],
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "66b8db97e4b437c6b2569e79"
                    },
                    "name": {
                        "type": "string",
                        "example": "customer"
                    },
                    "email": {
                        "type": "string",
                        "example": "customer@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "Customer Password"
                    },
                    "role": {
                        "type": "string",
                        "example": "Customer",
                        "enum": [
                            "Customer",
                            "Deliveryman"
                        ]
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}